requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.0.3
gunicorn==20.1.0  # Added Gunicorn
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
Flask==3.0.3
sqlite3
